backend/requirements.txt
fastapi
uvicorn
requests
pandas
biopython
pylitsense
